+++ clcache/__main__.py
@@ -9,6 +9,7 @@
 from collections import defaultdict, namedtuple
 from ctypes import windll, wintypes
 from shutil import copyfile, copyfileobj, rmtree, which
+import argparse
 import cProfile
 import codecs
 import concurrent.futures
@@ -1203,29 +1204,31 @@ class ArgumentT4(Argument):
 
 
 class CommandLineAnalyzer:
+    argumentsWithParameter = {
+        # /NAMEparameter
+        ArgumentT1('Ob'), ArgumentT1('Yl'), ArgumentT1('Zm'),
+        # /NAME[parameter]
+        ArgumentT2('doc'), ArgumentT2('FA'), ArgumentT2('FR'), ArgumentT2('Fr'),
+        ArgumentT2('Gs'), ArgumentT2('MP'), ArgumentT2('Yc'), ArgumentT2('Yu'),
+        ArgumentT2('Zp'), ArgumentT2('Fa'), ArgumentT2('Fd'), ArgumentT2('Fe'),
+        ArgumentT2('Fi'), ArgumentT2('Fm'), ArgumentT2('Fo'), ArgumentT2('Fp'),
+        ArgumentT2('Wv'),
+        # /NAME[ ]parameter
+        ArgumentT3('AI'), ArgumentT3('D'), ArgumentT3('Tc'), ArgumentT3('Tp'),
+        ArgumentT3('FI'), ArgumentT3('U'), ArgumentT3('I'), ArgumentT3('F'),
+        ArgumentT3('FU'), ArgumentT3('w1'), ArgumentT3('w2'), ArgumentT3('w3'),
+        ArgumentT3('w4'), ArgumentT3('wd'), ArgumentT3('we'), ArgumentT3('wo'),
+        ArgumentT3('V'),
+        ArgumentT3('imsvc'),
+        # /NAME parameter
+        ArgumentT4("Xclang"),
+    }
+    argumentsWithParameterSorted = sorted(argumentsWithParameter, key=len, reverse=True)
 
     @staticmethod
     def _getParameterizedArgumentType(cmdLineArgument):
-        argumentsWithParameter = {
-            # /NAMEparameter
-            ArgumentT1('Ob'), ArgumentT1('Yl'), ArgumentT1('Zm'),
-            # /NAME[parameter]
-            ArgumentT2('doc'), ArgumentT2('FA'), ArgumentT2('FR'), ArgumentT2('Fr'),
-            ArgumentT2('Gs'), ArgumentT2('MP'), ArgumentT2('Yc'), ArgumentT2('Yu'),
-            ArgumentT2('Zp'), ArgumentT2('Fa'), ArgumentT2('Fd'), ArgumentT2('Fe'),
-            ArgumentT2('Fi'), ArgumentT2('Fm'), ArgumentT2('Fo'), ArgumentT2('Fp'),
-            ArgumentT2('Wv'),
-            # /NAME[ ]parameter
-            ArgumentT3('AI'), ArgumentT3('D'), ArgumentT3('Tc'), ArgumentT3('Tp'),
-            ArgumentT3('FI'), ArgumentT3('U'), ArgumentT3('I'), ArgumentT3('F'),
-            ArgumentT3('FU'), ArgumentT3('w1'), ArgumentT3('w2'), ArgumentT3('w3'),
-            ArgumentT3('w4'), ArgumentT3('wd'), ArgumentT3('we'), ArgumentT3('wo'),
-            ArgumentT3('V'),
-            # /NAME parameter
-        }
         # Sort by length to handle prefixes
-        argumentsWithParameterSorted = sorted(argumentsWithParameter, key=len, reverse=True)
-        for arg in argumentsWithParameterSorted:
+        for arg in CommandLineAnalyzer.argumentsWithParameterSorted:
             if cmdLineArgument.startswith(arg.name, 1):
                 return arg
         return None
@@ -1527,46 +1530,78 @@ def createManifestEntry(manifestHash, includePaths):
 
 
 def main():
-    if len(sys.argv) == 2 and sys.argv[1] == "--help":
-        print("""
-clcache.py v{}
-  --help    : show this help
-  -s        : print cache statistics
-  -c        : clean cache
-  -C        : clear cache
-  -z        : reset cache statistics
-  -M <size> : set maximum cache size (in bytes)
-""".strip().format(VERSION))
-        return 0
+    # These Argparse Actions are necessary because the first commandline
+    # argument, the compiler executable path, is optional, and the argparse
+    # class does not support conditional selection of positional arguments.
+    # Therefore, these classes check the candidate path, and if it is not an
+    # executable, stores it in the namespace as a special variable, and
+    # the compiler argument Action then prepends it to its list of arguments
+    class CommandCheckAction(argparse.Action):
+        def __call__(self, parser, namespace, values, optional_string=None):
+            if values and not values.lower().endswith(".exe"):
+                setattr(namespace, "non_command", values)
+                return
+            setattr(namespace, self.dest, values)
+
+    class RemainderSetAction(argparse.Action):
+        def __call__(self, parser, namespace, values, optional_string=None):
+            nonCommand = getattr(namespace, "non_command", None)
+            if nonCommand:
+                values.insert(0, nonCommand)
+            setattr(namespace, self.dest, values)
+
+    parser = argparse.ArgumentParser(description="clcache.py v" + VERSION)
+    # Handle the clcache standalone actions, only one can be used at a time
+    groupParser = parser.add_mutually_exclusive_group()
+    groupParser.add_argument("-s", "--stats", dest="show_stats",
+                             action="store_true",
+                             help="print cache statistics")
+    groupParser.add_argument("-c", "--clean", dest="clean_cache",
+                             action="store_true", help="clean cache")
+    groupParser.add_argument("-C", "--clear", dest="clear_cache",
+                             action="store_true", help="clear cache")
+    groupParser.add_argument("-z", "--reset", dest="reset_stats",
+                             action="store_true",
+                             help="reset cache statistics")
+    groupParser.add_argument("-M", "--set-size", dest="cache_size", type=int,
+                             default=None,
+                             help="set maximum cache size (in bytes)")
+
+    # This argument need to be optional, or it will be required for the status commands above
+    parser.add_argument("compiler", default=None, action=CommandCheckAction,
+                        nargs="?",
+                        help="Optional path to compile executable. If not "
+                             "present look in CLCACHE_CL environment variable "
+                             "or search PATH for cl.exe.")
+    parser.add_argument("compiler_args", action=RemainderSetAction,
+                        nargs=argparse.REMAINDER,
+                        help="Arguments to the compiler")
+
+    options = parser.parse_args()
 
     cache = Cache()
 
-    if len(sys.argv) == 2 and sys.argv[1] == "-s":
+    if options.show_stats:
         printStatistics(cache)
         return 0
 
-    if len(sys.argv) == 2 and sys.argv[1] == "-c":
+    if options.clean_cache:
         cleanCache(cache)
         print('Cache cleaned')
         return 0
 
-    if len(sys.argv) == 2 and sys.argv[1] == "-C":
+    if options.clear_cache:
         clearCache(cache)
         print('Cache cleared')
         return 0
 
-    if len(sys.argv) == 2 and sys.argv[1] == "-z":
+    if options.reset_stats:
         resetStatistics(cache)
         print('Statistics reset')
         return 0
 
-    if len(sys.argv) == 3 and sys.argv[1] == "-M":
-        arg = sys.argv[2]
-        try:
-            maxSizeValue = int(arg)
-        except ValueError:
-            print("Given max size argument is not a valid integer: '{}'.".format(arg), file=sys.stderr)
-            return 1
+    if options.cache_size is not None:
+        maxSizeValue = options.cache_size
         if maxSizeValue < 1:
             print("Max size argument must be greater than 0.", file=sys.stderr)
             return 1
@@ -1575,18 +1610,19 @@ clcache.py v{}
             cfg.setMaximumCacheSize(maxSizeValue)
         return 0
 
-    compiler = findCompilerBinary()
-    if not compiler:
-        print("Failed to locate cl.exe on PATH (and CLCACHE_CL is not set), aborting.")
+
+    compiler = options.compiler or findCompilerBinary()
+    if not (compiler and os.access(compiler, os.F_OK)):
+        print("Failed to locate specified compiler, or cl.exe on PATH (and CLCACHE_CL is not set), aborting.")
         return 1
 
     printTraceStatement("Found real compiler binary at '{0!s}'".format(compiler))
     printTraceStatement("Arguments we care about: '{}'".format(sys.argv))
 
     if "CLCACHE_DISABLE" in os.environ:
-        return invokeRealCompiler(compiler, sys.argv[1:])[0]
+        return invokeRealCompiler(compiler, options.compiler_args)[0]
     try:
-        return processCompileRequest(cache, compiler, sys.argv)
+        return processCompileRequest(cache, compiler, options.compiler_args)
     except LogicException as e:
         print(e)
         return 1
@@ -1605,9 +1641,9 @@ def printErrStr(message):
         print(message, file=sys.stderr)
 
 def processCompileRequest(cache, compiler, args):
-    printTraceStatement("Parsing given commandline '{0!s}'".format(args[1:]))
+    printTraceStatement("Parsing given commandline '{0!s}'".format(args))
 
-    cmdLine, environment = extendCommandLineFromEnvironment(args[1:], os.environ)
+    cmdLine, environment = extendCommandLineFromEnvironment(args, os.environ)
     cmdLine = expandCommandLine(cmdLine)
     printTraceStatement("Expanded commandline '{0!s}'".format(cmdLine))
 
@@ -1638,7 +1674,7 @@ def processCompileRequest(cache, compiler, args):
         printTraceStatement("Cannot cache invocation as {}: called for preprocessing".format(cmdLine))
         updateCacheStatistics(cache, Statistics.registerCallForPreprocessing)
 
-    exitCode, out, err = invokeRealCompiler(compiler, args[1:])
+    exitCode, out, err = invokeRealCompiler(compiler, args)
     printOutAndErr(out, err)
     return exitCode