+++ CHANGELOG.md
@@ -8,6 +8,9 @@ clcache changelog
  * Bugfix: Fixed a race condition resulting in clcache to fail with an error
    message like `PermissionError: [WinError 5] Access is denied:
    'XXXX\\stats.txt.new' -> 'XXXX\\stats.txt'` (GH #334)
+ * The path to the compiler executable can optionally be specified on the
+   command line, instead of with an environment variable, or searching the PATH. 
+ * Added support for clang-cl
 
 ## clcache 4.2.0 (2018-09-06)